Folder structure Refactor and import changes
